Getter functions.

Functions to extract meta information of grapherator object.

getNumberOfNodes(graph)

getNumberOfEdges(graph)

getNumberOfClusters(graph)

getNumberOfWeights(graph)

getNodeCoordinates(graph, cluster.centers = FALSE)

getWeightMatrix(graph, objective)

getAdjacencyMatrix(graph)

getNodeDegrees(graph)

Arguments

graph
[grapherator]
Graph.

cluster.centers
[logical(1)]
Return coordinates of cluster centers only? Default is FALSE.

objective
[integer(1)]
Number of weight matrix to return.

Examples

g = graph(0, 100)
g = addNodes(g, n = 25, generator = addNodesGrid)
g = addEdges(g, generator = addEdgesGrid)
g = addWeights(g, generator = addWeightsRandom, method = runif, min = 5, max = 100, to.int = TRUE)
g = addWeights(g, generator = addWeightsDistance, method = "euclidean")

getNumberOfNodes(g)

#> [1] 25

getNumberOfEdges(g)

#> [1] 40

getNumberOfClusters(g)

#> [1] 0

getNumberOfWeights(g)

#> [1] 2

getNodeCoordinates(g)

#>       [,1] [,2]
#>  [1,]    0    0
#>  [2,]   25    0
#>  [3,]   50    0
#>  [4,]   75    0
#>  [5,]  100    0
#>  [6,]    0   25
#>  [7,]   25   25
#>  [8,]   50   25
#>  [9,]   75   25
#> [10,]  100   25
#> [11,]    0   50
#> [12,]   25   50
#> [13,]   50   50
#> [14,]   75   50
#> [15,]  100   50
#> [16,]    0   75
#> [17,]   25   75
#> [18,]   50   75
#> [19,]   75   75
#> [20,]  100   75
#> [21,]    0  100
#> [22,]   25  100
#> [23,]   50  100
#> [24,]   75  100
#> [25,]  100  100

getWeightMatrix(g, 2)

#>      1   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17  18  19
#> 1    0  25 Inf Inf Inf  25 Inf Inf Inf Inf Inf Inf Inf Inf Inf Inf Inf Inf Inf
#> 2   25   0  25 Inf Inf Inf  25 Inf Inf Inf Inf Inf Inf Inf Inf Inf Inf Inf Inf
#> 3  Inf  25   0  25 Inf Inf Inf  25 Inf Inf Inf Inf Inf Inf Inf Inf Inf Inf Inf
#> 4  Inf Inf  25   0  25 Inf Inf Inf  25 Inf Inf Inf Inf Inf Inf Inf Inf Inf Inf
#> 5  Inf Inf Inf  25   0 Inf Inf Inf Inf  25 Inf Inf Inf Inf Inf Inf Inf Inf Inf
#> 6   25 Inf Inf Inf Inf   0  25 Inf Inf Inf  25 Inf Inf Inf Inf Inf Inf Inf Inf
#> 7  Inf  25 Inf Inf Inf  25   0  25 Inf Inf Inf  25 Inf Inf Inf Inf Inf Inf Inf
#> 8  Inf Inf  25 Inf Inf Inf  25   0  25 Inf Inf Inf  25 Inf Inf Inf Inf Inf Inf
#> 9  Inf Inf Inf  25 Inf Inf Inf  25   0  25 Inf Inf Inf  25 Inf Inf Inf Inf Inf
#> 10 Inf Inf Inf Inf  25 Inf Inf Inf  25   0 Inf Inf Inf Inf  25 Inf Inf Inf Inf
#> 11 Inf Inf Inf Inf Inf  25 Inf Inf Inf Inf   0  25 Inf Inf Inf  25 Inf Inf Inf
#> 12 Inf Inf Inf Inf Inf Inf  25 Inf Inf Inf  25   0  25 Inf Inf Inf  25 Inf Inf
#> 13 Inf Inf Inf Inf Inf Inf Inf  25 Inf Inf Inf  25   0  25 Inf Inf Inf  25 Inf
#> 14 Inf Inf Inf Inf Inf Inf Inf Inf  25 Inf Inf Inf  25   0  25 Inf Inf Inf  25
#> 15 Inf Inf Inf Inf Inf Inf Inf Inf Inf  25 Inf Inf Inf  25   0 Inf Inf Inf Inf
#> 16 Inf Inf Inf Inf Inf Inf Inf Inf Inf Inf  25 Inf Inf Inf Inf   0  25 Inf Inf
#> 17 Inf Inf Inf Inf Inf Inf Inf Inf Inf Inf Inf  25 Inf Inf Inf  25   0  25 Inf
#> 18 Inf Inf Inf Inf Inf Inf Inf Inf Inf Inf Inf Inf  25 Inf Inf Inf  25   0  25
#> 19 Inf Inf Inf Inf Inf Inf Inf Inf Inf Inf Inf Inf Inf  25 Inf Inf Inf  25   0
#> 20 Inf Inf Inf Inf Inf Inf Inf Inf Inf Inf Inf Inf Inf Inf  25 Inf Inf Inf  25
#> 21 Inf Inf Inf Inf Inf Inf Inf Inf Inf Inf Inf Inf Inf Inf Inf  25 Inf Inf Inf
#> 22 Inf Inf Inf Inf Inf Inf Inf Inf Inf Inf Inf Inf Inf Inf Inf Inf  25 Inf Inf
#> 23 Inf Inf Inf Inf Inf Inf Inf Inf Inf Inf Inf Inf Inf Inf Inf Inf Inf  25 Inf
#> 24 Inf Inf Inf Inf Inf Inf Inf Inf Inf Inf Inf Inf Inf Inf Inf Inf Inf Inf  25
#> 25 Inf Inf Inf Inf Inf Inf Inf Inf Inf Inf Inf Inf Inf Inf Inf Inf Inf Inf Inf
#>     20  21  22  23  24  25
#> 1  Inf Inf Inf Inf Inf Inf
#> 2  Inf Inf Inf Inf Inf Inf
#> 3  Inf Inf Inf Inf Inf Inf
#> 4  Inf Inf Inf Inf Inf Inf
#> 5  Inf Inf Inf Inf Inf Inf
#> 6  Inf Inf Inf Inf Inf Inf
#> 7  Inf Inf Inf Inf Inf Inf
#> 8  Inf Inf Inf Inf Inf Inf
#> 9  Inf Inf Inf Inf Inf Inf
#> 10 Inf Inf Inf Inf Inf Inf
#> 11 Inf Inf Inf Inf Inf Inf
#> 12 Inf Inf Inf Inf Inf Inf
#> 13 Inf Inf Inf Inf Inf Inf
#> 14 Inf Inf Inf Inf Inf Inf
#> 15  25 Inf Inf Inf Inf Inf
#> 16 Inf  25 Inf Inf Inf Inf
#> 17 Inf Inf  25 Inf Inf Inf
#> 18 Inf Inf Inf  25 Inf Inf
#> 19  25 Inf Inf Inf  25 Inf
#> 20   0 Inf Inf Inf Inf  25
#> 21 Inf   0  25 Inf Inf Inf
#> 22 Inf  25   0  25 Inf Inf
#> 23 Inf Inf  25   0  25 Inf
#> 24 Inf Inf Inf  25   0  25
#> 25  25 Inf Inf Inf  25   0

getAdjacencyMatrix(g)

#>        [,1]  [,2]  [,3]  [,4]  [,5]  [,6]  [,7]  [,8]  [,9] [,10] [,11] [,12]
#>  [1,] FALSE  TRUE FALSE FALSE FALSE  TRUE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [2,]  TRUE FALSE  TRUE FALSE FALSE FALSE  TRUE FALSE FALSE FALSE FALSE FALSE
#>  [3,] FALSE  TRUE FALSE  TRUE FALSE FALSE FALSE  TRUE FALSE FALSE FALSE FALSE
#>  [4,] FALSE FALSE  TRUE FALSE  TRUE FALSE FALSE FALSE  TRUE FALSE FALSE FALSE
#>  [5,] FALSE FALSE FALSE  TRUE FALSE FALSE FALSE FALSE FALSE  TRUE FALSE FALSE
#>  [6,]  TRUE FALSE FALSE FALSE FALSE FALSE  TRUE FALSE FALSE FALSE  TRUE FALSE
#>  [7,] FALSE  TRUE FALSE FALSE FALSE  TRUE FALSE  TRUE FALSE FALSE FALSE  TRUE
#>  [8,] FALSE FALSE  TRUE FALSE FALSE FALSE  TRUE FALSE  TRUE FALSE FALSE FALSE
#>  [9,] FALSE FALSE FALSE  TRUE FALSE FALSE FALSE  TRUE FALSE  TRUE FALSE FALSE
#> [10,] FALSE FALSE FALSE FALSE  TRUE FALSE FALSE FALSE  TRUE FALSE FALSE FALSE
#> [11,] FALSE FALSE FALSE FALSE FALSE  TRUE FALSE FALSE FALSE FALSE FALSE  TRUE
#> [12,] FALSE FALSE FALSE FALSE FALSE FALSE  TRUE FALSE FALSE FALSE  TRUE FALSE
#> [13,] FALSE FALSE FALSE FALSE FALSE FALSE FALSE  TRUE FALSE FALSE FALSE  TRUE
#> [14,] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE  TRUE FALSE FALSE FALSE
#> [15,] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE  TRUE FALSE FALSE
#> [16,] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE  TRUE FALSE
#> [17,] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE  TRUE
#> [18,] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [19,] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [20,] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [21,] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [22,] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [23,] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [24,] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [25,] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>       [,13] [,14] [,15] [,16] [,17] [,18] [,19] [,20] [,21] [,22] [,23] [,24]
#>  [1,] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [2,] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [3,] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [4,] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [5,] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [6,] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [7,] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [8,]  TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [9,] FALSE  TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [10,] FALSE FALSE  TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [11,] FALSE FALSE FALSE  TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [12,]  TRUE FALSE FALSE FALSE  TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [13,] FALSE  TRUE FALSE FALSE FALSE  TRUE FALSE FALSE FALSE FALSE FALSE FALSE
#> [14,]  TRUE FALSE  TRUE FALSE FALSE FALSE  TRUE FALSE FALSE FALSE FALSE FALSE
#> [15,] FALSE  TRUE FALSE FALSE FALSE FALSE FALSE  TRUE FALSE FALSE FALSE FALSE
#> [16,] FALSE FALSE FALSE FALSE  TRUE FALSE FALSE FALSE  TRUE FALSE FALSE FALSE
#> [17,] FALSE FALSE FALSE  TRUE FALSE  TRUE FALSE FALSE FALSE  TRUE FALSE FALSE
#> [18,]  TRUE FALSE FALSE FALSE  TRUE FALSE  TRUE FALSE FALSE FALSE  TRUE FALSE
#> [19,] FALSE  TRUE FALSE FALSE FALSE  TRUE FALSE  TRUE FALSE FALSE FALSE  TRUE
#> [20,] FALSE FALSE  TRUE FALSE FALSE FALSE  TRUE FALSE FALSE FALSE FALSE FALSE
#> [21,] FALSE FALSE FALSE  TRUE FALSE FALSE FALSE FALSE FALSE  TRUE FALSE FALSE
#> [22,] FALSE FALSE FALSE FALSE  TRUE FALSE FALSE FALSE  TRUE FALSE  TRUE FALSE
#> [23,] FALSE FALSE FALSE FALSE FALSE  TRUE FALSE FALSE FALSE  TRUE FALSE  TRUE
#> [24,] FALSE FALSE FALSE FALSE FALSE FALSE  TRUE FALSE FALSE FALSE  TRUE FALSE
#> [25,] FALSE FALSE FALSE FALSE FALSE FALSE FALSE  TRUE FALSE FALSE FALSE  TRUE
#>       [,25]
#>  [1,] FALSE
#>  [2,] FALSE
#>  [3,] FALSE
#>  [4,] FALSE
#>  [5,] FALSE
#>  [6,] FALSE
#>  [7,] FALSE
#>  [8,] FALSE
#>  [9,] FALSE
#> [10,] FALSE
#> [11,] FALSE
#> [12,] FALSE
#> [13,] FALSE
#> [14,] FALSE
#> [15,] FALSE
#> [16,] FALSE
#> [17,] FALSE
#> [18,] FALSE
#> [19,] FALSE
#> [20,]  TRUE
#> [21,] FALSE
#> [22,] FALSE
#> [23,] FALSE
#> [24,]  TRUE
#> [25,] FALSE

getNodeDegrees(g)

#>  [1] 2 3 3 3 2 3 4 4 4 3 3 4 4 4 3 3 4 4 4 3 2 3 3 3 2