Graph object.

S3 object describing a graph with the following fields:

lower [numeric(2)]

Lower bounds for node coordinates in the Euclidean plane.

upper [numeric(2)]

Upper bounds for node coordinates in the Euclidean plane.

n.clusters [integer(1)]

Number of clusters.

n.nodes [integer(1)]

Number of nodes.

n.edges [integer(1)]

Number of edges.

n.weights [integer(1)]

Number of weights associated with each edge.

node.types [character]

Character vector describing the node generators used to create nodes.

edge.types [character]

Character vector describing the node generators used to create edges.

weight.types [character]

Character vector describing the node generators used to create weights.

weights [list of matrix]

List of weight/distance/cost matrizes.

degree [integer]

Integer vector of node degrees.

membership [integer | NULL]

Integer vector which stores the cluster membership of each node. Not NULL only if graph is clustered.

coordinates [matrix(n.nodes, 2)]

Matrix of node coordinates. Each row contains the node coordinates of one node.